+++ Assets/SoulRunProject/Scripts/OutGame/Interface/ITraitInterface.cs
