+++ Assets/SoulRunProject/Scripts/OutGame/Interface/Trait/TestTrait.cs
