+++ Assets/SoulRunProject/Scripts/OutGame/ScriptableObjects/SoulCardData.cs.meta
