+++ Assets/SoulRunProject/Scripts/OutGame/ScriptableObjects/TraitWrapper.cs
