--- Assets/SoulRunProject/Scripts/SoulMixScene/GameData.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7dc8802480adfc34f93025d7ea1264cd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
